requirements.txt
docker
git+https://github.com/hitomi-team/shimeji.git
ray[default]
discord.py
transformers
torch
dynaconf
mmappickle
bitsandbytes-cuda113